Move C++ code to separate header to be able to include MSCrashesPrivate.h to pure Objective-C file
