+++ AppCenterCrashes/AppCenterCrashesTests/MSCrashesTests.mm
@@ -2,11 +2,13 @@
 #import "MSAppleErrorLog.h"
 #import "MSChannelUnitConfiguration.h"
 #import "MSChannelUnitDefault.h"
+#import "MSCrashesBufferedLog.h"
 #import "MSCrashesInternal.h"
 #import "MSCrashesPrivate.h"
 #import "MSCrashesTestUtil.h"
 #import "MSCrashesUtil.h"
 #import "MSCrashHandlerSetupDelegate.h"
+#import "MSCrashReporter.h"
 #import "MSErrorAttachmentLogInternal.h"
 #import "MSErrorLogFormatter.h"
 #import "MSException.h"
